fix: rename get_XXX to describe_XXX for consistency
