+++ .secrets.baseline
@@ -713,6 +713,36 @@
         "is_secret": false
       }
     ],
+    "src/rds-control-plane-mcp-server/tests/resources/db_cluster/test_list_clusters.py": [
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "src/rds-control-plane-mcp-server/tests/resources/db_cluster/test_list_clusters.py",
+        "hashed_secret": "ee04099f4fb672975f0f5e75c056eb4768f40da0",
+        "is_verified": false,
+        "line_number": 110,
+        "is_secret": false
+      }
+    ],
+    "src/rds-control-plane-mcp-server/tests/resources/db_instance/test_describe_instance_detail.py": [
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "src/rds-control-plane-mcp-server/tests/resources/db_instance/test_describe_instance_detail.py",
+        "hashed_secret": "e988de1eb61164b4ba5db0f741f789892ea949cf",
+        "is_verified": false,
+        "line_number": 56,
+        "is_secret": false
+      }
+    ],
+    "src/rds-control-plane-mcp-server/tests/resources/db_instance/test_list_instances.py": [
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "src/rds-control-plane-mcp-server/tests/resources/db_instance/test_list_instances.py",
+        "hashed_secret": "e988de1eb61164b4ba5db0f741f789892ea949cf",
+        "is_verified": false,
+        "line_number": 98,
+        "is_secret": false
+      }
+    ],
     "src/stepfunctions-tool-mcp-server/README.md": [
       {
         "type": "Base64 High Entropy String",
@@ -750,5 +780,5 @@
       }
     ]
   },
-  "generated_at": "2025-07-03T06:19:26Z"
+  "generated_at": "2025-07-21T13:05:15Z"
 }
