+++ src/rds-control-plane-mcp-server/awslabs/rds_control_plane_mcp_server/resources/db_cluster/describe_cluster_detail.py
